requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
openai
requests
streamlit
httpx